Add FlashFinder
This is still very much a work in progress. It got too late and I needed to get to bed, so I haven't tested this.

FlashFinder checks dice for potential flashes, and stores them in a FlashBus. The reason for this approach is that there could be two potential flashes (assuming one Flaming Sun and then two like pairs), so there has to be a way to return multiple potential flashes at the same time.